Fix: Use tracefile size/count to prevent storing the trace
The --no-output flag is ignored in live mode, the traces are actually
stored by the relay in ~/lttng-traces/<hostname>/<session-name>. In
live mode we need some storage because the client asks the relay for
data and the relay cannot store the trace in memory for an arbitrary
amount of time.

A way to prevent the trace from always increasing in size is to use the
--tracefile-size and --tracefile-count options to limit the size of the
tracefiles on disk, the relay will automatically overwrite older files a
keep a ring-buffer of tracefiles on disk. If the client is too slow to
fetch the data before it is overwritten, we will see lost packets, so we
need to increase the amount we keep.

In this patch, we keep at most 5 files of 1M of trace for each CPU, if
it is not enough, it will be easy to change.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>